+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Balanced Dashboard Changelog
 
+### 1.0.3
+
+* Fixed images in import payouts page.
+* Fixed duplicate "Logs" tab in search results
+* Fixed csv import flow.
+* Fixed funding instrument title in disputes page.
+
 ### 1.0.2
 
 * Fixed bug where new marketplaces would not link to the user after being created.